:recycle: Move train helpers to train_utils
